+++ CHANGELOG.md
@@ -9,7 +9,15 @@
 - New API `NBTCodec#writeTag(ByteBuffer, Tag)` for writing NBT tags to a byte buffer.
 - New API `NBTCodec#writeTagToByteArray(Tag)` for writing NBT tags to a byte array.
 - New API `ArrayTag#getElementType()` for getting the element type of array tags.
+- New API `NBTSchema` for validating NBT elements.
 
+### Breaking Changes
+
+- The `NBTValidator` interface has been rewritten and now provides two methods for validation:
+    - [#test(NBTElement)]: Returns `true` if the element meets the rules of this validator, otherwise returns `false`.
+    - [#validate(NBTElement)]: Does nothing if the element meets the rules of this validator, otherwise throws an
+      exception.
+  
 ## 0.1.0 (2026-03-16)
 
 - Initial release